app/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.3
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.61.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-extensions==3.2.3
django-leaflet==0.29.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.6
oauthlib==3.2.2
packaging==23.2
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==2.2.1
